requirements.txt
streamlit
pandas
openpyxl
plotly
openai
python-dotenv
fpdf2